+++ packages/api-client/src/views/Request/RequestRoot.test.ts
@@ -4,6 +4,7 @@ import { useActiveEntities } from '@/store/active-entities'
 import { mount } from '@vue/test-utils'
 import { type Mock, beforeEach, describe, expect, it, vi } from 'vitest'
 import { ref } from 'vue'
+import { mockUseLayout } from '@/vitest.setup'
 
 import RequestRoot from './RequestRoot.vue'
 
@@ -88,5 +89,12 @@ describe('RequestRoot', () => {
     const wrapper = createWrapper()
     await wrapper.find('.scalar-sidebar-toggle').trigger('click')
     expect(wrapper.find('.scalar-sidebar-toggle').exists()).toBe(true)
+  })
+
+  it('applies correct classes for modal layout', async () => {
+    mockUseLayout.mockReturnValue({ layout: 'modal' })
+    const wrapper = createWrapper()
+    const sidebarToggle = wrapper.find('.scalar-sidebar-toggle')
+    expect(sidebarToggle.classes()).toContain('!flex')
   })
 })
